test: migrate MediaWikiAPIs tests in source
